Backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
django-environ==0.9.0
django-storages[google]==1.13
google-cloud-secret-manager==2.12.0
psycopg2-binary==2.9.4
gunicorn==20.1.0
whitenoise==6.3.0
django-cors-headers==3.13.0
numpy==1.24.2